# swagger_openapi_saludvital.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión Médica - Salud Vital API
  description: |
    API REST para el sistema de gestión médica integral. 
    Permite la gestión de pacientes, médicos, citas médicas y recetas.
    
    ## Usuarios Predefinidos para Pruebas:
    
    ### 👑 **ADMIN**
    - **Email**: `admin@saludvital.com`
    - **Password**: `admin123`
    
    ### 🩺 **MÉDICO** 
    - **Email**: `medico@saludvital.com`
    - **Password**: `medico123`
    
    ### 👤 **PACIENTE**
    - **Email**: `paciente@saludvital.com`
    - **Password**: `paciente123`
    
    ## Flujo de Autenticación:
    1. Hacer login con alguno de los usuarios predefinidos
    2. Copiar el `accessToken` de la respuesta
    3. Usar en header: `Authorization: Bearer {accessToken}`
  version: 1.0.0
  contact:
    name: Equipo de Desarrollo
    email: dev@saludvital.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api
    description: Servidor de Desarrollo
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer Token. Ejemplo 'Bearer {token}'
  schemas:
...
    CitaRequest:
      type: object
      required:
        - medicoId
        - fecha
        - hora
      properties:
        medicoId:
          type: integer
          format: int64
          example: 1
        fecha:
          type: string
          format: date
          example: "2024-12-20"
        hora:
          type: string
          format: time
          example: "10:00:00"
        motivo:
          type: string
          example: "Consulta general"
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operación exitosa"
        data:
          type: object
...
paths:
...
  /paciente/perfil:
    get:
      tags:
        - 👤 Pacientes
      summary: Ver perfil del paciente
      description: Obtiene la información del perfil del paciente autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: No autenticado
        '403':
          description: Sin permisos (solo pacientes)
  /paciente/expediente:
    get:
      tags:
        - 👤 Pacientes
      summary: Ver expediente médico
      description: Obtiene el expediente médico completo del paciente
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Expediente obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /medico/perfil:
    get:
      tags:
        - 🩺 Médicos
      summary: Ver perfil del médico
      description: Obtiene la información del perfil del médico autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Perfil obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo médicos)
  /medico/{id}:
    get:
      tags:
        - 🩺 Médicos
      summary: Obtener médico por ID
      description: Obtiene la información de un médico específico
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Médico obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: Médico no encontrado
  /citas/nueva:
    post:
      tags:
        - 📅 Citas
      summary: Solicitar nueva cita
      description: |
        Crear una nueva cita médica (solo pacientes).
        
        **Validaciones:**
        - Máximo 3 citas por paciente por día
        - No horarios duplicados para el mismo médico
        - Solo en horarios de atención del médico
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitaRequest'
      responses:
        '200':
          description: Cita creada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Error de validación (máximo 3 citas, horario ocupado, etc.)
        '403':
          description: Sin permisos (solo pacientes)
  /citas/mis-citas:
    get:
      tags:
        - 📅 Citas
      summary: Ver mis citas
      description: Obtiene las citas del paciente autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Citas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /citas/medico:
    get:
      tags:
        - 📅 Citas
      summary: Ver citas del médico
      description: Obtiene las citas del médico autenticado
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Citas obtenidas exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo médicos)
  /citas/{id}:
    put:
      tags:
        - 📅 Citas
      summary: Modificar cita
      description: |
        Actualiza una cita existente.
        
        **Validación:** Modificación con al menos 2 horas de anticipación.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitaRequest'
      responses:
        '200':
          description: Cita actualizada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: No se puede modificar (menos de 2 horas de anticipación)
  /citas/{id}/cancelar:
    put:
      tags:
        - 📅 Citas
      summary: Cancelar cita
      description: |
        Cancela una cita existente.
        
        **Validación:** Cancelación con al menos 2 horas de anticipación.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Cita cancelada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /citas/{id}/completar:
    put:
      tags:
        - 📅 Citas
      summary: Completar cita
      description: Marca una cita como completada (solo médicos)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Cita completada exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '403':
          description: Sin permisos (solo médicos)
...